Run tests on iPhone 16 Pro
This is an arbitrary change, but I noticed that if I run the tests on
my machine with iPhone 16, the Simulator fails to launch:

```
Test target AnalyticsTests encountered an error (Failed to clone device named 'iPhone 16'. (Underlying Error: The operation couldn’t be completed. Device was allocated but was stuck in creation state.  Check CoreSimulator.log for more information.))
```

I tried to run the tests via Fastlane's `scan` and it automatically
picked up iPhone 16 Pro and succeeded to run the tests.

When I changed the command to use iPhone 16 Pro on my end via vanilla
`xcodebuild` like CI does, it worked.